+++ app/services/weeks/start_service.rb
@@ -10,12 +10,14 @@ module Weeks
       price_change_service: Teams::Players::CorrectPriceService,
       cup_create_service: Cups::CreateService,
       cups_pairs_generate_service: Cups::Pairs::GenerateService,
-      generate_week_position: GENERATE_WEEK_POSITION
+      generate_week_position: GENERATE_WEEK_POSITION,
+      refresh_selected_by_teams: FantasySports::Container['services.players.seasons.refresh_selected']
     )
       @price_change_service = price_change_service
       @cup_create_service = cup_create_service
       @cups_pairs_generate_service = cups_pairs_generate_service
       @generate_week_position = generate_week_position
+      @refresh_selected_by_teams = refresh_selected_by_teams
     end
 
     def call(week:)
@@ -25,6 +27,7 @@ module Weeks
       # commento: weeks.status
       week.update!(status: Week::ACTIVE)
       change_price_for_teams_players(week)
+      @refresh_selected_by_teams.call(season_id: week.season_id)
       generate_cups(week)
       generate_cups_pairs(week)
       add_week_fantasy_league(week)