+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - table view for squad views
+- precalculating of selected by teams ratio for players
 
 ## [1.1.4] - 2023-11-30
 ### Added